backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-bigquery==3.21.0
google-auth==2.29.0
pydantic==2.7.1
alembic==1.13.1
python-multipart==0.0.9
openai>=1.40.0
